Putting in a script I've been working on to automate generating the Obara-Saika recursion evaluation code (currently the files under "qmmm_terms" and "coulomb_terms")
The script can generate code for evaluation of arbitrary
2- 3- or 4- index integrals using the Obara-Saika rules up to
f basis functions, as well as gradients

It might help optimization on the existing QM/MM and Coulomb
code, as well as future work in the Coulomb energy
integrals, full 4-index integrals, or extending to f functions.

The script mostly works, but the gradient code it outputs still
needs some manual editing for now.