QM/MM AND COULOMB KERNELS NOW USE DOUBLE PRECISION BY DEFAULT
The single/mixed precision scheme (calculating prefactors with
double precision and doing the rest of the kernel in single precision)
which was the default before hasn't been fully validated on long
simulations for big systems, so to be safe double is now the default

Added compiler flags to turn on mixed precision separately for the
QM/MM (qmmm_mp) and Coulomb (coulomb_mp) kernels.  These flags
turn on mixed precision only for that kernel, and have no effect,
for instance, on the XC code.  The QM/MM and Coulomb kernels should
also still obey the full_double option