-To make it easier to persist GPU data for the QM/MM and Coulomb calculations, I'm restructuring that part of the code to be more modular with classes, etc.
-Separated the QM/MM and Coulomb code into a separate "analytic_integral"
directory.  For now, this code still gets built as part of g2g in
terms of Makefiles, but eventually it should be easy to completely
separate it with its own Makefile (maybe its own library, as well?)

-For right now, the QM/MM and Coulomb calculatations share a lot of common
data/code, so mixed precision can't be specified for each of them
individually anymore; rather, mixed precision can be turned on for
both now with the aint_mp flag. This also meant I had to take out a small
optimization to the QM/MM forces code, slowing it down a little,
until I get a change to reimplement for both QM/MM and Coulomb.

-Mixed precision seems to work OK now for the Coulomb fitting calculation;
however, convergence seems to take an extra step or two (sometimes more
for the initial density)