Bugfix + New Subroutines (#102)
* (1) Fixed a bug in the makefiles that was causing to not recompile modified
    files inside modules:

 *  cublasmath/cublasmath.mk
 *  general_module/general_module.mk
 *  linear_algebra/linear_algebra.mk
 *  liomods/liomods.mk
 *  maskrmm/maskrmm.mk
 *  mathsubs/mathsubs.mk

(2) Added a few subroutines for later use:

 *  Makefile.depends
    (to compile with new modules)
 *  errlog.f90
 *  lio_keywords.f90
 *  liosubs/catch_error.f90
 *  liosubs/find_free_unit.f90
    (modified slightly)
 *  liosubs/liosubs.f90
    (to include new subs)
 *  liosubs/liosubs.mk
    (to include new subs + previous fix)
 *  liosubs/safeio_open.f90
 *  liosubs/safeio_rewind.f90

* Small change: now the file, module and namelist liokeys are all named
the same. Didn't do this in the begining because I was afraid there
might be some conflict if a namelist and module had the same name, but
changed my mind. I'm keeping it this way for simplicity and change it
when and if such conflict is detected.