Overlap PR3.5 - Cleanups by hiding CUBLAS (#178)
First steps towards hiding cublas switch properly: added mask_cublas
module and introduced subroutines there that take care of initialization
if cublas is enabled or do nothing if it is not. This is a first step,
not the final result: there are still some ifdefs in SCF that we need
to eliminate.

Left as a separated PR to make it stand out from the rest of the cleanup.