(#253) Bugfix - dgesdd call in int2
I was getting a segmentation fault in this call that came and go with random modifications like adding calls to print (with a hardcoded text, not even variables). Problem was (apparently) that dgesdd was being called with a hardcoded number in places where it was expecting an array. Supposedly, because of other parameters it shouldn't be using those arrays, but seems like it needed them to be arrays anyway.