(#364) Add a Github action to run a code linter
Adds a Github action that will autocheck code-style in order to have a consistent style across the codebase.

This won't affect any forks since it will only run on develop and master branches, so people are free to code as they see fit, but will get a warning if the code doesn't pass the check when they try to merge upstream.

For now all languages are disabled (cpp, python, bash, latex) except for yaml in order to not have spurious warnings. Other PRs will format those accordingly and then those languages can be enabled. A fortran linter is not available in this docker image, but might be available in others, so we can try to find one to enable.