CMAKE IMPLEMENTATION FOR LIO
DESCRIPTION
The new CMake implementation is designed to work alongside the old makefile implementation until it seen fit to abandon it at some time in the future. The main goal of the CMake implementation is not so much as to facilitate compilation (it is not any easyier or difficult than before, I believe) but to make the compilation process more transparent and particularly, easier to understand and mantain. The makefile implentation consisted of about 41 makefiles spread about the whole code tree and bearing as complex an interdependence as the code files themselves. All this complexity is hidden now as CMake can, by itself, determine the interdependence of the code files and therefore the order in which they should be compiled, which means that the CMake compilation process is controlled by no more than 12 CMakeLists.txt files (only 4 of which are actually in charge of the compilation of the main code, with the rest belonging to testing and tools compilation). The implementation requires CMake version 3.18 or newer, which is a conservative requirement as Ubuntu 22.04 comes with CMake 3.22 and updating it to an even newer version is very easy to do.
CMake compilation is mainly controlled by files named CMakeLists.txt. The implementation consists of one root CMakeLists.txt file, three main ones (one for each of the three main parts of the code: g2g, lioamber and liosolo), a cmake directory containing modules (anciliary files that are normally included into main CMakeLists.txt files) and CMakeLists.txt files for testing and tools compilation.
The new CMake compilation option scheme, although preserves the original makefile implementation functionality, differs in its presentation. The new options scheme is more in tune with current CMake good practices, and are thought out to have more descriptive names. A set of previously undocumented options was also added to the main roster alongside comments describing its rough functionality as best I could devise them. The complete set of lio-specific compilation options is defined in the file cmake/ConfigureOptions.cmake. Also the defaults are set identically to the original makefile implementation.
MINOR CODE CHANGES
A new build_info_cmake.f90 was created as an update of the old build_info.f90. The main task of this file is to log compilation information from the compilation options, so naturally, it had to be recreated to reflect the new CMake option scheme.
A new lio_setenv_cmake.sh was added at the root directory that replaces the old liohome.sh script in the case of CMake compilation.
The directory tree for the test cases was changed to better automatize its execution. The energy python script in test/energy.py was refactored and updated to make it possible for the user to change the error threshold defaults (which were hardcoded in the original code) as user input to better control the checks precision goals depending of the case. This change was motivated by some tests systematically failing although the differences were actually within acceptable tolerances.
Added some comments con g2g/Makefile.
Eliminated option for compiler flags for extremely old intel compilers as those are unlikely to ever be used again.
Added some comments and caution marks on some undocumented options. In particular fastmath option is allways activated when compiling, but in a way that appears to be uningtentional. Added a comment to highlight this fact. In the CMake implementation this option is part of standard options and documented accordingly. Also comented the undocumented variable subscript_warn.

BUGFIXES
g2g/init.cpp: Added a precompiler definition to allow for compilation without OpenMP.
g2g/regenerate_partition.cpp: Modified to allow for serial compilation (no OpenMP). Changed a print which called an OpenMP function.
lioamber/TD.f90: Corrected logic of if() block. CMake compilation does not stop evaluation after (td_eu_step /= 0) is evaluated as false, and causes crash when mod(istep, td_eu_step) is undefined. Corrected the if block making it nested which is a safer option.
lioamber/cdft/cdft_utils.f90: Changed "Iteration n°:" for "Iteration No:" as the character ° was not compatible with utf-8 necoding and caused errors in the python scripts in charge of checking test runs.
lioamber/converger_subs/converger_ls.f90: Similar changes to cdft_utils.f90.
lioamber/fileio/output_scf.f90: Changes in the output format. Changed F12.6 for F14.6 as in some tests energy values would overflow.
lioamber/liomain.f90: Changed unsafe if() statement logic. Similar case to that of TD.f90. The evaluation of if ((restart_freq > 0) .and. (MOD(npas, restart_freq) == 0))
did not stop before the "and" if restart_freq=0 and produced a NAN when evaluating the MOD function causing lio to crash. It was rewritten as a nested if which is safer.
lioamber/properties/write_population.f90: Similar changes to cdft_utils.f90.
test/Makefile: Modified the makefile to reflect the new directory names and script names for tests.
test/run_test_external_functionals.py: Modified the script to reflect changes in the directory tree.